requirements.txt
# MySQL MCP Server 依赖项
python-mcp>=0.1.0
mysql-connector-python>=8.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.0.0
pytest-cov>=4.1.0